re-upload with new arrangement for auto checking
