+++ _docommand.c
@@ -2,28 +2,28 @@
 
 /**
  * _docommand - do a command
- * @argumentvectors: vector of arguments
- * Return: 0 on success
+ * @status: int, status
+ * @arguments: vector of arguments
+ * Return: int, status
  */
-int _docommand(char **argumentvectors)
+int _docommand(char **arguments, int status)
 {
 	pid_t pid;
-	int i = 0;
 
-	if (access(argumentvectors[i], F_OK) == 0)
+	pid = fork();
+	if (pid < 0)
 	{
-		pid = fork();
-		if (pid != 0)
-		{
-			wait(NULL);
-		}
-		if (pid == 0)
-		{
-			execve(argumentvectors[i], argumentvectors, NULL);
-		}
-		return (0);
+		perror("Could not create child process :(");
+		return (status);
 	}
-	if (access(argumentvectors[i], F_OK != 0))
-		_printf("%s: No such file or directory\n", argumentvectors[i]);
-	return (0);
+
+	if (pid == 0)
+		if (execve(arguments[0], arguments, NULL) == -1)
+			perror("Could not execute command");
+	if (pid > 0)
+	{
+		wait(&status);
+		status = WEXITSTATUS(status);
+	}
+	return (status);
 }
