+++ main.h
@@ -23,8 +23,11 @@ int _strncmp(char *s1, char *s2, int n);
 int _strlen(char *s);
 int _prompt(void);
 char **_vectorize(char *buffer);
-int _docommand(char **argumentvectors);
+int _docommand(char **arguments, int status);
 char *_strdup(char *str);
 int _freeargs(char **arguments);
+int _interpret_commands(int status, char *buffer);
+char *str_concat(char *s1, char *s2);
+char *_look_for_a_program(char *name);
 
 #endif /* MAIN_H */
