--- newshell/AUTHORS
@@ -1,4 +1,0 @@
-# This file contains those who helped contribute to this shell project
-
-Douglas Davison <4939@holbertonstudents.com>
-Caleb House <4922@holbertonstudents.com>
