--- newshell/testfile
@@ -1,3 +1,0 @@
-/usr/bin/ls
-boulder
-/usr/bin/ls
